requirements.txt
# Core dependencies for ICU Deterioration Prediction

# Deep Learning
torch>=2.0.0
transformers>=4.30.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0
polars>=0.18.0

# Machine Learning
scikit-learn>=1.2.0

# Configuration
pyyaml>=6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars
tqdm>=4.65.0

# Logging
tensorboard>=2.13.0

# Utilities
python-dateutil>=2.8.0
